🔨 Refactor: 아이 엔티티 필드 타입 수정
- 몸무게와 동일하게 키도 실수형으로 수정

+++ src/main/java/final_project/momeasy/domain/child/entity/Child.java
@@ -33,7 +33,7 @@ public class Child extends BaseEntity {
     private LocalDate birthdate;
 
     @Column(nullable = false)
-    private int height;
+    private float height;
 
     @Column(nullable = false)
     private float weight;
